tests/wpt/web-platform-tests/html/interaction/focus/focusgroup/tentative/backward-navigation/does-not-move-when-only-one-item.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focusgroup - Focus does not move when there is only one item.</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com/">
<link rel="help" href="https://github.com/MicrosoftEdge/MSEdgeExplainers/blob/main/Focusgroup/explainer.md">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="../resources/focusgroup-utils.js"></script>

<div focusgroup>
  <span id=item1 tabindex=0>item1</span>
</div>

<script>

  promise_test(async t => {
    var item1 = document.getElementById("item1");

    await focusAndKeyPress(item1, kArrowUp);
    assert_equals(document.activeElement, item1);

    await focusAndKeyPress(item1, kArrowLeft);
    assert_equals(document.activeElement, item1);
  }, "When the focus is set on the only focusgroup item, the focus shouldn't move and we shouldn't get stuck in an infinite loop.");

</script>
